requirements.txt
# Core ML / NLP
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# Data processing
pandas>=2.1.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
pdfminer.six>=20221105
lxml>=4.9.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
